Implement message controller.
This patchset contributes a message controller for Marconi client. It
implements simple property querying for messages, as well as the
ability to reload these properties as needed (polling?) and delete a
message. It also uses the assumed structure of the HTTPclient that
will form the basis of the connection from the common libraries.

six is used to boost py3k compat.

This also introduces a directory structure for supporting multiple
transports. Instead of having a 'controller' dir, now we have a
'transports/http' and 'transports/zmq'.

Implements: blueprint messages-management
Change-Id: I287583ebff69c4571ae8164c99cff3b0628d23a1